Consolidate architecture and implement MCP server with @rekog/mcp-nest
Major architectural consolidation and MCP implementation:
- Consolidate libs/temporal-bridge-core into apps/temporal-bridge-cli/src/lib/
- Implement MCP server using @rekog/mcp-nest v1.8.2 with STDIO transport
- Add mcp.ts entrypoint alongside main.ts with full NestJS integration
- Implement 15+ MCP tools with @Tool decorators and Zod schema validation
- Fix dependency injection issues that caused IDE "no usages" warnings
- Update all imports to use relative paths instead of package references
- Remove cross-package complexity while maintaining all functionality
- Update CLAUDE.md with PNPM usage and current architecture
- Update conversion plan docs with consolidation rationale

Tools implemented: search_personal, search_all, get_recent_episodes,
get_current_context, share_knowledge, list_projects, project_context,
project_technologies, get_technology_expertise, get_thread_context
