+++ apps/temporal-bridge-cli/package.json
@@ -9,15 +9,21 @@
   "scripts": {
     "build": "tsc",
     "dev": "tsx src/main.ts",
+    "dev:mcp": "tsx src/mcp.ts",
     "start": "node dist/main.js",
+    "start:mcp": "node dist/mcp.js",
     "typecheck": "tsc --noEmit"
   },
   "dependencies": {
-    "@temporal-bridge/core": "workspace:*",
+    "@getzep/zep-cloud": "^3.2.0",
+    "@rekog/mcp-nest": "^1.8.2",
     "nest-commander": "^3.13.0",
     "@nestjs/common": "^10.4.4",
     "@nestjs/core": "^10.4.4",
-    "reflect-metadata": "^0.2.2"
+    "@nestjs/platform-express": "^10.4.4",
+    "reflect-metadata": "^0.2.2",
+    "yaml": "^2.6.1",
+    "zod": "^3.23.8"
   },
   "devDependencies": {
     "@types/node": "^22.7.7",
