+++ apps/temporal-bridge-cli/src/lib/project-detector.ts
