+++ apps/temporal-bridge-cli/src/lib/session-manager.ts
