+++ apps/temporal-bridge-cli/src/lib/zep-client.ts
@@ -8,7 +8,7 @@ import { Injectable } from '@nestjs/common';
 export { ZepError };
 
 // Re-export common Zep types
-export type Reranker = 'rrf' | 'mmr' | 'node_distance' | 'episode_mentions' | 'cross_encoder';
+export type Reranker = 'rrf' | 'mmr' | 'node_distance' | 'episode_mentions' | 'cross_encoder' | 'none';
 
 interface ApiError extends Error {
   statusCode?: number;
